notes from PR, squash before merge
